Rename `edge-config-fs` to `global-config-fs`
