+++ .github/workflows/deprecate-package.yml
@@ -11,6 +11,7 @@ on:
           - "@vercel/blob"
           - "@vercel/edge-config"
           - "@vercel/edge-config-fs"
+          - "@vercel/global-config-fs"
           - "@vercel/global-config"
           - "@vercel/kv"
           - "@vercel/postgres"
