--- packages/edge-config-fs/README.md
@@ -1,4 +1,0 @@
-# `@vercel/edge-config-fs`
-
-This is an internal package in support of `@vercel/edge-config`.
-You should always use `@vercel/edge-config` directly.
