+++ packages/global-config-fs/.eslintrc.js
