+++ packages/global-config-fs/README.md
@@ -1,0 +1,4 @@
+# `@vercel/global-config-fs`
+
+This is an internal package in support of `@vercel/global-config`.
+You should always use `@vercel/global-config` directly.
