+++ packages/global-config-fs/index.edge.cjs
