+++ packages/global-config-fs/index.edge.js
