+++ packages/global-config-fs/index.node.cjs
