+++ packages/global-config-fs/index.node.js
